fixed update_license.py

+++ bim_project/license_activation/update_license.py
@@ -16,21 +16,49 @@ pwd = os.getcwd()
 
 
 
+def define_purpose():
+    
+    purpose = input("\nCheck the licenses(1) || Apply license(2) || Delete active license(3) || Exit(q)\nSelect what to do: ").lower()
+    print()
+    if purpose in ("1", "check",):
+        return 'check'
+    elif purpose in ("2", "update"):
+        return 'update'
+    elif purpose in ("3", "delete"):
+        return 'delete'
+    else:
+        sys.exit("Exit.")
+
+def get_data_from_license_file():
+
+    data_from_lic_file: dict = {}
+
+    # Check if the *.lic file is in place
+    if os.path.isfile(f"{pwd}/license.lic"):
+
+        with open("license.lic", "r", encoding="utf-8") as file:    # get license_id from the .lic file and put it into data_from_lic_file dictionary
+            for x in file.read().split('\n'):
+                if 'LicenseID' in x.strip():
+                    data_from_lic_file['LicenseID'] = x.split()[2]
+                    break
+
+        with open("license.lic", "r", encoding="utf-8") as file:    # get license_token from the .lic file and put it into data_from_lic_file dictionary
+            data_from_lic_file['license_token'] = file.read().split()[0].strip("\"")
+            
+        return data_from_lic_file
+    
+    else:
+        sys.exit("No .lic file in the folder. Exit.")
+
+
 
 def creds():
 
     logging.basicConfig(filename=f"{pwd}/license_log.txt", level=logging.DEBUG,
                         format="%(asctime)s %(levelname)s - %(message)s", filemode="w", datefmt='%d-%b-%y %H:%M:%S')
 
-    # Check if the *.lic file is in place
-    if os.path.isfile(f"{pwd}/license.lic"):    
-        with open("license.lic", "r", encoding="utf-8") as file:        
-            license_token = file.read().split()[0].strip("\"")
-    else:
-        sys.exit("No .lic file in the folder. Exit.")
 
-
-    url = input("Enter URL: ").lower()
+    url = input("\nEnter URL: ").lower()
     if url[-1:] == '/':
         url = url[:-1]
     else: pass
@@ -40,10 +68,8 @@ def creds():
     data_for_connect: dict = {
                     "url": url,
                     "username": username,
-                    "password": password,
-                    "license_token": license_token
+                    "password": password
                   }
-    print()
     return data_for_connect
 
 
@@ -55,9 +81,12 @@ def get_token():
     headers = {'accept': '*/*', 'Content-type':'application/json; charset=utf-8'}
     url_get_providers = f"{data_for_connect['url']}/api/Providers"
 
-    
-    id_request = requests.get(url_get_providers, headers=headers, verify=False)
-    response = id_request.json() 
+    try:
+        id_request = requests.get(url_get_providers, headers=headers, verify=False)
+        response = id_request.json()
+    except requests.exceptions.MissingSchema as err:
+        logging.error('%s', err)
+        sys.exit(f"Connection error: {err}")
 
     for dict in response:
         list_of_providersID.append(dict['id'])
@@ -71,18 +100,40 @@ def get_token():
                     "providerId": id
                 }
         data = json.dumps(payload)
-    
-        auth_request = requests.post(url_auth, data=data, headers=headers, verify=False)
-        response = auth_request.json()
 
-        if auth_request.status_code == 200:
+        
+        auth_request = requests.post(url_auth, data=data, headers=headers, verify=False)
+        response = auth_request.json()       
+
+        if auth_request.status_code in (200, 201, 204):
             token = response['access_token']
             break
         else:
             logging.error("%s", auth_request.status_code)
-            continue
+            sys.exit(f"Connection error: {auth_request.status_code}")
 
     return token
+
+
+
+'''   Show all the licenses in the system   '''
+def show_licenses():
+    print()
+    url = f"{data_for_connect['url']}/api/License"
+    headers = {'Content-type':'text/plane', 'Authorization': f"Bearer {token}"}
+    payload = {
+                "username": data_for_connect['username'],
+                "password": data_for_connect['password']
+              }
+    
+    request = requests.get(url, data=payload, headers=headers, verify=False)
+    request.raise_for_status()
+
+    # response is a list of dictionaries with a set of keys: 'isActive', 'serverId', 'licenseID', 'until', 'activeUsers', 'activeUsersLimit'
+    response = request.json()
+    for license in response:
+        print(license)
+    print()
 
 
 '''   Get the list of licenses   '''
@@ -100,8 +151,9 @@ def get_license():
 
     # response is a list of dictionaries with a set of keys: 'isActive', 'serverId', 'licenseID', 'until', 'activeUsers', 'activeUsersLimit'
     response = request.json()
-    return response
 
+    return response
+    
 
 
 '''   Delete active license, if there is one.   '''
@@ -109,15 +161,25 @@ def delete_license():
 
     headers = {'accept': '*/*', 'Content-type': 'text/plane', 'Authorization': f"Bearer {token}"}
 
+    # Check if there are any active licenses
+    count = 0
     for dct in get_license():
-        if dct['isActive'] == True and dct['licenseID'] != '00000000-0000-0000-0000-000000000000':
-            url = f"{data_for_connect['url']}/api/License/{dct['licenseID']}" 
-            request = requests.delete(url, headers=headers, verify=False)
-            # response = request.text
-            if request.status_code == 204:
-                print(f"License has been deactivated: {dct['licenseID']}")
-            else:
-                print(request.status_code)
+        if dct.get('isActive'):
+            count += 1
+    
+    if count == 0:
+        print("No active licenses have been found in the system.")
+    else:
+        for license in get_license():
+            if license['isActive'] == True and license['licenseID'] != '00000000-0000-0000-0000-000000000000':
+                url = f"{data_for_connect['url']}/api/License/{license['licenseID']}" 
+                request = requests.delete(url, headers=headers, verify=False)
+                if request.status_code in (200, 201, 204,):
+                    print(f"License '{license['licenseID']}' has been deactivated!")
+                else:
+                    logging.error('%s', request.text)
+                    print(f"Error with deactivation: {request.status_code}")
+            
 
 
 
@@ -126,12 +188,14 @@ def post_license():
     headers = {'accept': 'text/plane', 'Content-Type': 'application/json-patch+json', 'Authorization': f"Bearer {token}"}
     url = f"{data_for_connect['url']}/api/License/"
 
-    data = json.dumps(data_for_connect['license_token'])
+    license_token = get_data_from_license_file()['license_token']
+    data = json.dumps(license_token)
 
     request = requests.post(url, headers=headers, data=data, verify=False)
-    if request.status_code == 200:
+    if request.status_code in (200, 201, 204,):
         print("License has been posted!")
     else:
+        logging.error('%s', request.text)
         print("License has not been posted!")
 
 
@@ -140,28 +204,43 @@ def put_license():
     
     headers = {'accept': '*/*', 'Content-type': 'text/plane', 'Authorization': f"Bearer {token}"}
 
-    # {'isActive': True, 'serverId': '6e71c225-bb65-4ff7-b831-bc0d1894396b', 'licenseID': '309d6c41-1f4b-41b2-938d-b2069e54b12a', 'until': '2022-08-25T23:59:59', 'activeUsers': 1, 'activeUsersLimit': 1500}
-    for dct in get_license():
-        if dct['isActive'] == True:
-            url = f"{data_for_connect['url']}/api/License/active/{dct['licenseID']}"
-            data = json.dumps('no_need_to_put_anything')
-            request = requests.put(url, headers=headers, data=data, verify=False)
-            if request.status_code == 204:
-                print(f"License has been activated: {dct['licenseID']}")
-            else:
-                print(f"License has not been activated: {dct['licenseID']}")            
-            break
+    license_id = get_data_from_license_file()['LicenseID']
+    
+    url = f"{data_for_connect['url']}/api/License/active/{license_id}"
+    data = json.dumps('no_need_to_put_anything')
+    request = requests.put(url, headers=headers, data=data, verify=False)
+    if request.status_code in (200, 201, 204,):
+        print(f"License '{license_id}' has been activated.")
+    else:
+        logging.error('%s', request.text)
+        print(f"License '{license_id}' has not been activated.")        
+        print()
+            
+            
 
     
 
-# http://192.168.87.66
-# http://std-p7.bimeister.io
-# https://bim-class.bimeister.io
-# Qwerty12345!
+
+'''
+http://192.168.87.66
+http://std-p7.bimeister.io
+https://bim-class.bimeister.io
+Qwerty12345!
+'''
 
 if __name__ == "__main__":
     data_for_connect = creds()
     token = get_token()
-    delete_license()
-    post_license()
-    put_license()
+    check_or_update = define_purpose()
+    if check_or_update == 'update':
+        post_license()
+        put_license()
+    elif check_or_update == 'check':
+        pass
+    elif check_or_update == 'delete':
+        delete_license()
+
+    show_licenses()
+    if sys.platform == 'win32':
+            os.system('pause')
+
